requirements.txt
python-dotenv==1.0.1
requests==2.32.2
python-telegram-bot==13.0.0
retry==0.9.2
